requirements-image-dedupe.txt
Pillow>=10.0.0
ImageHash>=4.3.1
send2trash>=1.8.2
